requirements.txt
openai>=1.0.0
requests>=2.28.0
schedule>=1.2.0
numpy>=1.21.0
pandas>=1.3.0
fastapi>=0.95.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
jinja2>=3.1.0
alpinejs>=3.0.0
tailwindcss>=3.0.0
